requirements.txt
tensorflow
keras
numpy
matplotlib
scikit-learn
pillow